Complete user deletion

+++ src/BeautySalonManager/bsm.dal/Repositories/UserRepository.cs
@@ -28,14 +28,20 @@ namespace bsm.dal.Repositories
 
         public void AddRow(User user)
         {
-            _context.Users.Add(user);
-            _context.SaveChanges();
+            if(user != null)
+            {
+                _context.Users.Add(user);
+                _context.SaveChanges();
+            }
         }
 
         public void DeleteRow(User user)
         {
-            _context.Users.Remove(user);
-            _context.SaveChanges();
+            if(user != null)
+            {
+                _context.Users.Remove(user);
+                _context.SaveChanges();
+            }
         }
     }
 }
